The user can add a complete line
Lots of theater are composed of a lines of seats.
To avoid the user to insert its seats one by one,
we offer it to insert a whole line.

By `line`, it means a row or a column.

To accelerate things at first, the user inputs use `prompt`.
Further, we should create a proper modal system.

This commit does the following:

- Create some utilities
- Create new action `AddLine`
- Implement the whole lifecycle of this new action
- Adapt the alignment line to consider this new case