The user can add a complete grid
Lots of theater are composed of a grid of seats.
To avoid the user to insert its seats one by one (or line by line),
we offer it to insert a whole grid.

The grid can have a shift like we see in lots of theaters.

To accelerate things at first, the user inputs use `prompt`.
Further, we should create a proper modal system.

This commit does the following:

- Add a new toolbar button
- When activated, create the grid for the `addingSeats` action