Correct behavior of seat selection and moving
With recent updates, we change the behavior of seat selection.
Instead of simply clicking on each seat to select/deselect them,
we need to press `ctrl` to add a seat in the selection otherwise the selection reset with the new seat.

We also remove the shortcut `ctrl+click` to rename a seat.
The right panel is enough for that and avoid strange behavior.

This commit does the following:

- Change `toggleSelectSeat` to `selectSeat` and its behavior (see above)
- The click on `SeatElement` select the seat, with or without addition to the others
- Always show the moving shadow (no remaining issue)