Add the notion of stage
As an information, we want to be able to indicate a stage in the map.
A stage can have a different size of a seat but all the remaining information are the same.

The implemented solution was to consider a Stage as a type of Seat,
but its type is not bookable.

This solution is user overridable but is also a simple and permissive solution.

This commit does the following:

- Add attribute `Type.bookable` to define a bookable Seat/Stage
- Add attributes `Seat.width` and `Seat.height`
- Consider those attributes in the rest of the code (instead of constants `seatWidth` and `seatHeight`)
- Update the default `defaultStructure` with a type of seats and a type of stages