Add `width` and `height` attributes for preview
Sometimes, we want to manage the size of the map.
If we force `width` and `height` from the size map, it will never be resizable.

The map can be resized from the container or from the `width` and `height` attributes.
Also, to provide more control, the user can manage the `preserveAspectRatio`.

This commit does the following:

- Add `width`, `height` and `preserveAspectRatio` in `MapPreview`
- Manage those attribute across the project (web-component and preact)
- Add samples with container size and attribute sizes